ngx-live: support filler content
adds the ability to set filler audio/video content, that will be used in
case no audio/video is received.
when only some audio/video tracks are missing, the gap is filled from
other tracks. the filler is used only when all tracks of a certain media
type are missing.