ngx-live: skip non-key-frames before segment start
even though ngx_stream_live_kmp_module waits for the first key frame,
the segmenter may flush its pending frames in case of inactivity,
without the stream module being aware of it.
the stream mechanism for skipping non-key-frames is left as is, in order
to avoid sending non-key-frames to the syncer.