ngx-rtmp-kmp: avoid republish when nothing to send
when audio frames stop arriving, ngx-live closes the connection once the
idle timer expires. in this case, there's no reason to perform
republish, since it will just timeout again in a loop.
instead, if the kmp connection drops and there's nothing to send, wait
with the republish until more data is available.