ngx-mods: improve json reader
make it more symmetrical with json writer - use a python script to
generate a header file with a hash table for mapping the json props.
the read process also decodes the string (unless explicitly marked as
'raw') and makes sure the same key does not repeat more than once.

also renamed "preset_name" in channel json to "preset" to make it
symmetrical with channel.add API.