ngx-rtmp-kmp: invalid connect packet type fix (#607)
must not append buffers to u->last when the kmp upstream is not
connected and the connect packet wasn't sent. if the buffers are
appended in this state, the connect packet may be appended after them.

in order to make sure it doesn't happen, removed 'sent_buffered' and
instead using u->last != NULL - u->last is not null only when the kmp
upstream is connected and the connect packet was sent.