ngx-live: check track id in media info pending add
after a channel is restored from persistence, the pending queue is
empty, even if a track used another track as source (this can't happen
unless the channel is restored, since 'set source' saves the last media
info to the pending queue).
when a new media info came in via kmp and it matched the last media info
node exactly, it wasn't added. however, the last node pointed to another
track. this resulted in incorrect track ids in the media info nodes,
possibly resulting in failure to find segment data in serve media reqs.

the fix is to compare the track id of the last node in 'pending add', if
it doesn't match the track, the media info must be added.
also, added an if to make sure a node belonging to another track will
never be added to the pending queue in 'set source' (it won't be used
anyway, so no reason to add it there)