ngx-pckg-dash: add basic signaling of 608 captions
this is a very minimal impl - if all tracks report cea captions, add the
accessibility tag to the mpd.
a more complete implementation would be to add the ability to pass a
mapping of caption id/lang/label to nginx live, and use it here.
as was done here for vod -
https://github.com/kaltura/nginx-vod-module/pull/1228