ngx-pckg: add pckg_media_type_selector directive
when set to request (default) media type selectors of index/segment urls
inside returned manifests will be set according to the selectors sent on
the request.
when set to accual, the selectors will be set according to the available
media types. for example, if there are no video tracks on the channel,
returned index/segment urls will have a suffix of `-a`.