ngx-live: change copy track selection logic
do not require has_last_segment, but prefer a track that is not a
filler.
for example, if the previous segment used a filler for video, we still
want to use a regular (non-filler) track, even though its
has_last_segment is false. the reason is that the filler track won't
have the video content that was previously published.

also - added a missing call to ngx_rbtree_insert in
ngx_live_media_info_queue_copy.