ngx-live: segment info read bug fix
following 88e4dc01a77d904cd364c9dad22d0e33a42a5231, the first element in
a delta file may be skipped, however code assumed that all elements will
be read, and it set nelts accordingly.
the fix is not to assume the number is known in advance, and continue as
long as there are more elements in both src & dst. after the loop, the
number of elements is set according to the actual number of elements
written.