ngx-live: avoid delta persist after history change
currently, a change to past segments can occur in 2 scenarios -
1. a track is added - its media info + segment info data is copied from
   another track
2. a timeline is copied from another timeline

in these case, must not use delta persistence, if delta is used, the
file may not be readable (timeline may not find the initial segment
index) or it may be missing data (media info/segment info)

it is possible that the next saved snapshot after the change will not
include the newly added track/timeline yet, forcing it to a full index
will not suffice, the next delta will break. to solve this one, a full
index save is forced until it includes the segment in which the history
change occurred.