ngx-live: manifest timeline persist fix
the manifest timeline fields are set when the index is written, not
when the snapshot is taken. this was done intentionally, since the
timeline can advance between these two points in time - the manifest
timeline position when the manifest is taken, can be outside (too far
behind) the timeline at the time of writing the index.

however, this implementation can lead to the opposite scenario in which
the manifest timeline advances so much between the snapshot time and
index write time, that it's outside the range (too far ahead) of the
timeline. this is because the timeline is clipped to the max_index at
the time the snapshot was taken.

the solution is to calculate the manifest timeline fields for the last
segment in the timeline when taking the snapshot. if the manifest
timeline moves ahead of the timeline position during snapshot, the
values saved in the snapshot will be used. if the saved timeline is not
empty, it must contain the last segment of the snapshot, so these values
will always be valid.

usually, the manifest timeline will not move ahead of the timeline
position, and in these cases, the values during the index write will be
used (as in the previous implementation)