ngx-pckg: expose subrequest variables on main
upstream vars evaluate only on the subrequest, in order to make them
available on the main request, we define these vars with a pckg_ prefix
and proxy them to the subrequest for evaluation.

also added copy of variables on the subrequest creation - the default
behavior in nginx is to share all vars between the main request and all
subrequests, and this could lead to unexpected behavior (e.g. after
evaluating $pckg_upstream_status, $upstream_status will return the same
value, because it was set by the subrequest)