ngx-live: handle large inter-frame jumps
seems some encoders have timestamp wraparound issues causing them to
generate large pts jumps even within gop. such jumps must be removed,
otherwise, the resulting segment may have a duration that is too large.
an extremely large segment can push the 'last pts' too far into the
future, preventing the module from generating additional segments for a
long time.

the solution is to add another threshold (100 sec by default), if this
threshold is exceeded, a sync will be performed even inside a gop.
in addition, within the gop, need to make sure the resulting dts will be
larger than the previous one, to avoid a negative frame duration.

also - postponed the track inactive event to when the track loses all
frames, so that the syncer state will be reset only then, and not when
the segmenter inactive timeout fires. reason - the segmenter inactive
timer may fire, but a segment will not be closed immediately because
some other track is pending. if some frame arrives in this state, we
should use the previous correction value, instead of resynching.