ngx-live: syncer reconnect fix
after kmp disconnect/reconnect, all pending frames are disposed by the
segmenter module (ngx_live_segmenter_remove_all_frames).
in this case, the syncer state should be reverted to what it was on the
last (non-disposed) frame. if the state is not reverted, last_pts may be
too much ahead, and the syncer will "see" a backward jump in pts.
if the jump is large (>10 sec) the syncer would resync and cause a
disconnection on the stream, even though no real jump in pts has
occurred.
the solution is to use track->last_frame_pts which holds the pts of the
last non-disposed frame, and add to it a similar field for dts.
the syncer can use these 2 fields + its change log to reconstruct the
state.