ngx-live-ll: dispose after choosing pending index
move the decision whether to dispose a frame to be after the selection
of pending index. the previous impl can be wrong if there is a large gap
in the timing of pending indexes, and a track joins late. in this case,
we may get a segment that is too long due to frames that fall in the gap