ngx-pckg: swap date-time/map tags
swap the order of EXT-X-MAP & EXT-X-PROGRAM-DATE-TIME in order to work
around hls.js bug -
https://github.com/video-dev/hls.js/commit/a4b4f5f1f0f5cc9d6c5199be28e4edb1cddabf91

when the date time is before the init segment, hls.js sets its value on
the init segment and from there propagates it to the segments. due to
the bug, the same date time value propagates to ALL segments.
this makes the player load the wrong segment when trying to match the
audio/video content, delaying playback start.