ngx-kmp-push: add connect_data field to upstream
the connect_data field can be returned on upstream objects in
publish/republish responses. its value must be a base64-encoded string,
the value is decoded and sent as the data of the KMP_CONNECT packet.
when a republish response has the connect_data field, it overrides any
previously set connect_data. if the republish response does not contain
connect_data, the previously set value will be used.