ngx-live: support forced end_list
when playing the recording timeline of an active entry, we want to
return EXT-X-ENDLIST in order to -
1. make the player start from the beginning
2. freeze the stream so that it can be clipped

when using low latency mode, there are always pending segments when the
timeline is active, so the end_list setting is ignored.

the solution is to upgrade end_list from a bool to enum - "off", "on",
"forced". when the value is "forced", the pending segments are excluded
and EXT-X-ENDLIST is returned.