ngx-live: add input_delay param
used for delaying the closing of segments in the segmenter, to give
captions that arrive late (e.g. human captioning) a chance to make it
into the segment.
the input_delay is measured based on the KMP 'created' timestamp of the
frame. this behavior was chosen instead of using the clock, so that in
case of reconnect, frames that already waited long enough, will be
processed immediately.
since the 'created' timestamp may have been assigned by a different
server, due to clock differences, it is possible that the frame's
created value is in the future. if the value is slightly ahead, it's
assumed to be valid (controlled by the input_delay_margin conf param).
if it's too far ahead, it's considered invalid, and the module will
process the frame immediately.
the input_delay param is configurable via nginx.conf & API.