ngx-ts-kmp: send publish only after connect done
currently, a failed connect request terminates the incoming connection.
but the module does not wait for connect to complete before sending the
publish request.
the implication of this is that resources may be allocated for a stream,
and data may be published, when the url does not have a valid token.

the solution is to upgrade the "error" flag of ngx_ts_kmp_ctx_t to an
enum with 3 states: initial, connect done, error.
the publish request is sent only when:
1. the connect request finished successfully
2. the media info of the track was received