ngx-kmp-out: last_written bug fix
when the frames that are written to the kmp-out-track are written in
parts (=using the frame_start / frame_end funcs), append is called
before the frame header was written (frame_start writes a zero frame
header, and frame_end updates it)
the 'written' value of the track includes the partial frame data that
was already written, so ngx_kmp_out_upstream_auto_ack tries to parse it.
however, since the frame header is zero, ngx_kmp_out_upstream_ack_packet
sees packet_type = 0 and fails.
the solution is to keep the 'written' value when the last frame ended,
and auto ack frames up to this value. this way, the header of the
partial frame will not be read, and no error will occur.