ngx-kmp-out: protect against slow upstream
when the kmp-out-upstream is connected, the auto ack is limited by the
number of sent bytes. if the upstream server is receiving very slowly,
this may prevent the "watermark" mechanism from freeing buffers.
eventually, the memory limit will be hit, and the client rtmp/mpegts
connection will be dropped.
note that if the upstream is not receiving at all, the write timeout
will kill the connection, the memory limit may be hit only if the
upstream server is receiving, but slowly.
the solution is to perform a republish in this case, the republish
closes the upstream connection, making it possible to acknowledge
packets that were not sent yet.