ngx-live: avoid close failed error
when capturing a frame from a segment that was persisted, and the
requested track is not found, nginx-live-module returns an empty
segment.
since no data is written, ngx_persist_write_close returns a NULL chain,
this makes the request fail with 'internal server error'.
the solution is set write_ctx to null when getting an empty segment
response.