ngx-live: avoid sending inactive events for subs
unlike video/audio, subtitles are not expected to be continuous - it is
possible that subtitles cues will be distributed sparsely.
when an inactive event is sent, the syncer module resets itself, and
outputs several log lines.
the main reason for avoiding the inactive event for subtitles is to
remove this clutter from the log.