ngx-live: fix audio ends before video issue
the segmenter may try to cut a segment at the audio end timestamp, there
is no video key frame at this timestamp - the video segments come out
empty, and some of the audio frames are discarded.
then it tries to create another segment, but the segments of all tracks
come out empty, and the channel is failed.
the fix is to remove all pending frames in instead of failing, if all
tracks are inactive