ngx-live: fix audio ends before video issue
the segmenter may try to cut a segment at the audio end timestamp, there
is no video key frame at this timestamp - the video segments come out
empty, and some of the audio frames are discarded. then it tries to
create another segment, but the segments of all tracks come out empty,
and the channel is failed.
the fix is to avoid averaging the min/max split pts when the span is too
large. in this case, the target pts will be the audio end pts, the video
segments will come out empty, and the trailing audio will be discarded.